+++ app/forms/views/submission.py
@@ -54,7 +54,9 @@ class SubmissionViewSet(APIFormErrorsMixin, BaseViewSet):
         if isinstance(form, EventForm):
             user = request.user
             event = form.event
-            attended = event.get_queue().filter(user=user, has_attended=True).exists()
+            attended = (
+                event.get_participants().filter(user=user, has_attended=True).exists()
+            )
 
             if form.type == EventFormType.EVALUATION and not attended:
                 return Response(